refactor(container): extract authentication handler providers
